% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{apply_code}
\alias{apply_code}
\title{Recoding and cleaning of survey data}
\usage{
apply_code(df, dict, oldVar, newVar)
}
\arguments{
\item{df}{data frame to change}

\item{dict}{Assumes that the dictionary will be a 2-column data frame with key in first column and value in the second. Third column is an optional column explaining the relationship.  Should be some sort of logical operator, e.g. '==', '<', '<=', ...}

\item{oldVar}{name of the variable to change}

\item{newVar}{name of the new variable to generate}
}
\description{
Converts values of a data frame to encoded values using a dictionary.
}
\examples{
df = data.frame(region_code = sample(6, 10, replace = TRUE))
dict = data.frame(region_code = 1:6, region_name = c('Africa', 'Asia', 'Australia', 'Europe', 'North America', 'South America'))

# straight dictionary
llamar::apply_code(df, dict, 'region_code', 'name')

# fancy math
dict2 = data.frame(region_code = 1:5, 
region_name = c('Africa', 'Asia', 'Australia', 'Europe', 'Americas'),
logic = c(rep('==', 4), '>=')) # If >= 5, assign to Americas.
llamar::apply_code(df, dict2, 'region_code', 'name')
}
