% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{attr}
\alias{attr}
\alias{pullAttributes}
\alias{removeAttributes}
\title{Remove or pull attributes (e.g. labels)}
\usage{
pullAttributes(data)

removeAttributes(data)
}
\arguments{
\item{data}{imported data frame}

\item{type}{for pullAttributes, type of metadata to pull (label or labels)}
}
\description{
Often when Stata .dta files are imported into R with the haven package, they contain variable and value labels.  
While these metadata are useful, if they are incomplete, they often have annoying downstream repercussions.
For instance, ggplot can't facet if there are value labels.
}
\details{
These functions allow you to save all labels and levels from a data frame, and to remove them.
}
\section{Functions}{
\itemize{
\item \code{pullAttributes}: Function to pull any attributes from a data frame.  Useful to cleanup imported Stata files.

\item \code{removeAttributes}: Function to remove any attributes from a data frame.  Useful to cleanup imported Stata files.
}}

