% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPtEst.R
\name{calcPtEst}
\alias{calcPtEst}
\title{Calculate point estimates and errors}
\usage{
calcPtEst(df, var, by_var = NA, design = NA, psu_var = NA,
  strata_var = NA, weight_var = NA, omit_NA = TRUE, ci_factor = 2)
}
\arguments{
\item{df}{main data frame containing the raw data}

\item{var}{string containing the variable name you want to average}

\item{by_var}{(optional) string containing the variable name over which you want to calculate the estimates}
}
\description{
Calculates point estimates and errors using the package `survey` based 
on sampling weights from a survey design
}
\author{
Laura Hughes, laura.d.hughes@gmail.com
}

