% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{category10}
\alias{category10}
\title{Colors: internal colors based on Vega and Tableau colors
From https://github.com/vega/vega/wiki/Scales}
\format{An object of class \code{character} of length 10.}
\usage{
category10
}
\description{
Colors: internal colors based on Vega and Tableau colors
From https://github.com/vega/vega/wiki/Scales
}
\keyword{datasets}
