% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.R
\name{coefplot}
\alias{coefplot}
\title{Plot coefficients of a model with errors}
\usage{
coefplot(model, level = 0.95, exclude_intercept = TRUE, size_point = 3)
}
\arguments{
\item{model}{Fitted model}
}
\description{
Inspired by https://github.com/jaredlander/coefplot/blob/master/R/coefplot.r
but works w/ ggplot2 version > 2.2
}
\examples{
data(diamonds)
model1 <- lm(price ~ carat + cut*color, data=diamonds)
coefplot(model1)

}

