% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert01.R
\name{convert01}
\alias{convert01}
\title{Recoding and cleaning of survey data}
\usage{
convert01(df, varName, newVar)
}
\arguments{
\item{varName}{variable to convert 1's and 2's to 1's and 0's}

\item{newVar}{variable name of new vari}
}
\description{
Function to convert survey data where 2 is no and 1 is yes.
}
\examples{
df = data.frame(x = c(2, 1, 9, NA), y = sample(100, 4))

df \%>\% convert01(varName = 'x', newVar = 'x2')

}

