% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize.R
\name{factorize}
\alias{factorize}
\title{Create factor levels based on labels from a dataset}
\usage{
factorize(df, ref_df, var, new_var)
}
\arguments{
\item{df}{main data frame containing the dataset to modify}

\item{ref_df}{reference labelled data frame whose labels you want to use}

\item{var}{string containing the variable name you want to change to a factor}

\item{new_var}{string for the name of the new factor-ized variable you create.}
}
\description{
Reads labels from a reference dataset (usually imported from Stata, SPSS, or SAS)
and applies them to the same variable in a label-free dataset to create a factor variable
with levels and labels. By default, the levels will be sorted by their frequency in the dataset.
}
\author{
Laura Hughes, laura.d.hughes@gmail.com
}

