% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_heatmap.R
\name{fcs_heatmap}
\alias{fcs_heatmap}
\title{Plot food consumption scores by region}
\usage{
fcs_heatmap(df, map_base = grey15K, map_accent = "#d53e4f", filename = NA,
  width = NA, height = NA, units = "in", scale = 1)
}
\description{
In its full version, a plot with four parts: (1) a heatmap breaking down the consumption of food
groups by region; (2) a kernel density estimate surface of the food consumption scores (FCS) of
individual households in the region; (3) a 1 x n heatmap of the average FCS score by region;
and (4) individual mini-maps highlighting the region for context.
}

