% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatNum.R
\name{formatNum}
\alias{formatNum}
\alias{percent}
\alias{roundMean}
\alias{roundStd}
\title{Simple reusable functions to format numbers}
\usage{
roundMean(x, ndigits = 2)

roundStd(x, ndigits = 2)

percent(x, ndigits = 1)
}
\arguments{
\item{x}{numbers to be converted}

\item{ndigits}{number of digits to which to round the number}
}
\description{
Simple reusable functions to format numbers
}
\section{Functions}{
\itemize{
\item \code{roundMean}: Function to calculate the mean and round to two digits.

\item \code{roundStd}: Function to calculate the standard deviation and round to two digits.

\item \code{percent}: Function to convert a decimal to a percent, round to a specific number of digits, and convert to a string with a percent sign.
}}
\examples{
roundMean(mtcars$mpg)
roundStd(mtcars$mpg)
percent(0.157389)
}

