% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDHScountry.R
\name{getDHScountry}
\alias{getDHScountry}
\title{Create a lookup table for DHS country codes, and find the country code based on the list}
\usage{
getDHScountry(country_list)
}
\description{
Can take either a single country or multiple; does
}

