% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_batch_free}
\alias{llama_batch_free}
\title{Free a llama batch allocated with \code{llama_batch_init()}}
\usage{
llama_batch_free(batch)
}
\arguments{
\item{batch}{An external pointer returned by \code{llama_batch_init()}.}
}
\value{
\code{NULL} invisibly.
}
\description{
Free a llama batch allocated with \code{llama_batch_init()}
}
\examples{
\dontrun{
batch <- llama_batch_init(512L)
llama_batch_free(batch)
}
}
