% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_chat_builtin_templates}
\alias{llama_chat_builtin_templates}
\title{List built-in chat templates}
\usage{
llama_chat_builtin_templates()
}
\value{
A character vector of built-in template names.
}
\description{
Returns a character vector of all chat template names supported by llama.cpp.
}
\examples{
# See which chat template formats are supported out of the box
templates <- llama_chat_builtin_templates()
head(templates)
}
