% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_max_devices}
\alias{llama_max_devices}
\title{Get maximum number of devices}
\usage{
llama_max_devices()
}
\value{
An integer scalar: the maximum number of compute devices available.
}
\description{
Get maximum number of devices
}
\examples{
# Query the maximum number of devices supported by the backend
n <- llama_max_devices()
cat("Max devices:", n, "\n")
}
