% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtmd.R
\name{llama_mtmd_support_audio}
\alias{llama_mtmd_support_audio}
\title{Does this multimodal context support audio?}
\usage{
llama_mtmd_support_audio(mctx)
}
\arguments{
\item{mctx}{An mtmd context from \code{\link{llama_mtmd_load}}.}
}
\value{
Logical scalar.
}
\description{
Does this multimodal context support audio?
}
