% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtmd.R
\name{llama_mtmd_support_vision}
\alias{llama_mtmd_support_vision}
\title{Does this multimodal context support vision (images)?}
\usage{
llama_mtmd_support_vision(mctx)
}
\arguments{
\item{mctx}{An mtmd context from \code{\link{llama_mtmd_load}}.}
}
\value{
Logical scalar.
}
\description{
Does this multimodal context support vision (images)?
}
