% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_n_batch}
\alias{llama_n_batch}
\title{Get logical batch size}
\usage{
llama_n_batch(ctx)
}
\arguments{
\item{ctx}{Context handle returned by [llama_new_context]}
}
\value{
An integer scalar: the logical batch size (max tokens per `llama_decode` call).
}
\description{
Get logical batch size
}
