% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_n_ctx_seq}
\alias{llama_n_ctx_seq}
\title{Get per-sequence context window size}
\usage{
llama_n_ctx_seq(ctx)
}
\arguments{
\item{ctx}{Context handle returned by [llama_new_context]}
}
\value{
An integer scalar: maximum context size per sequence.
}
\description{
Get per-sequence context window size
}
