% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_n_seq_max}
\alias{llama_n_seq_max}
\title{Get maximum number of sequences}
\usage{
llama_n_seq_max(ctx)
}
\arguments{
\item{ctx}{Context handle returned by [llama_new_context]}
}
\value{
An integer scalar: maximum number of concurrent sequences.
}
\description{
Get maximum number of sequences
}
