% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_n_threads}
\alias{llama_n_threads}
\title{Get number of threads for single-token generation}
\usage{
llama_n_threads(ctx)
}
\arguments{
\item{ctx}{Context handle returned by [llama_new_context]}
}
\value{
An integer scalar: current thread count for generation.
}
\description{
Get number of threads for single-token generation
}
