% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_n_threads_batch}
\alias{llama_n_threads_batch}
\title{Get number of threads for batch processing}
\usage{
llama_n_threads_batch(ctx)
}
\arguments{
\item{ctx}{Context handle returned by [llama_new_context]}
}
\value{
An integer scalar: current thread count for prompt encoding.
}
\description{
Get number of threads for batch processing
}
