% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_n_ubatch}
\alias{llama_n_ubatch}
\title{Get physical micro-batch size}
\usage{
llama_n_ubatch(ctx)
}
\arguments{
\item{ctx}{Context handle returned by [llama_new_context]}
}
\value{
An integer scalar: the physical micro-batch size.
}
\description{
Get physical micro-batch size
}
