% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_pooling_type}
\alias{llama_pooling_type}
\title{Get pooling type}
\usage{
llama_pooling_type(ctx)
}
\arguments{
\item{ctx}{Context handle returned by [llama_new_context]}
}
\value{
A character string: one of `"none"`, `"mean"`, `"cls"`, `"last"`,
  `"rank"`, `"unspecified"`.
}
\description{
Get pooling type
}
