% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_supports_mlock}
\alias{llama_supports_mlock}
\title{Check whether memory locking is supported}
\usage{
llama_supports_mlock()
}
\value{
A logical scalar: \code{TRUE} if mlock is supported.
}
\description{
Check whether memory locking is supported
}
\examples{
llama_supports_mlock()
}
