% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_vocab_is_control}
\alias{llama_vocab_is_control}
\title{Check if a token is a control token}
\usage{
llama_vocab_is_control(model, token)
}
\arguments{
\item{model}{Model handle returned by [llama_load_model]}

\item{token}{Integer token ID (0-based)}
}
\value{
A logical scalar.
}
\description{
Check if a token is a control token
}
