% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_vocab_type}
\alias{llama_vocab_type}
\title{Get vocabulary type}
\usage{
llama_vocab_type(model)
}
\arguments{
\item{model}{Model handle returned by [llama_load_model]}
}
\value{
A character string: one of `"spm"` (LLaMA/SentencePiece BPE),
  `"bpe"` (GPT-2 BPE), `"wpm"` (BERT WordPiece), `"ugm"` (T5 Unigram),
  `"rwkv"`, `"plamo2"`, or `"none"`.
}
\description{
Get vocabulary type
}
