% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_colour_text.R
\name{map_colour_text}
\alias{map_colour_text}
\title{Modifies a data frame to determine color to overlay on a colored background}
\usage{
map_colour_text(df, bckgrnd_column, colour_palette, limits = c(0,
  max(df[[bckgrnd_column]])), sat_threshold = 0.5, dark_colour = grey90K,
  light_colour = "white")
}
\arguments{
\item{df}{data frame containing the data}

\item{bckgrnd_column}{string containing the name of the column to map to fill values}

\item{colour_palette}{colour palette specification (list of hex values). Can use `RColorBrewer::brewer.pal` to generate}

\item{limits}{(optional) limits for the fill color palette mapping}

\item{sat_threshold}{(optional) breakpoint between the light and dark text color. 50% saturation, by default}

\item{dark_colour}{(optional) dark color to overlay on low fill values}

\item{light_colour}{(optional) light color to overlay on high fill values}
}
\description{
Takes a data frame with a value to map to a fill colour and determines whether
light or dark text should be used as the label on top of the fill. For use with
`ggplot2::scale_colour_identity()` downstream.
}
\examples{
# Define a Color Brewer palette
library(RColorBrewer)
# Generate random data
df = data.frame(x = 1:9, y = 1:9)
pal = 'Reds'

limits = c(0,15)
df = map_colour_text(df, 'x',  brewer.pal(9, pal), limits)

library(ggplot2)
ggplot(df, aes(x = x, y = y, fill = x, colour = text_colour, label = round(hsv.s,2))) +
geom_point(size = 10, shape = 21) + 
geom_text() +
scale_fill_gradientn(colours = brewer.pal(9, pal), limits = limits) +
scale_colour_identity() +
theme_blank()

}

