% tinyrox says don't edit this manually, but it can't stop you!
\name{memory_stats}
\alias{memory_stats}
\title{Get memory index statistics}
\usage{
memory_stats(agent_id = "default")
}
\arguments{
\item{agent_id}{Agent identifier}
}
\value{
List with file counts, chunk counts, etc.
}
\description{
Get memory index statistics
}
