% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{MODEL_CONTEXT_LIMITS}
\alias{MODEL_CONTEXT_LIMITS}
\title{Model context limits in tokens.}
\format{
An object of class \code{list} of length 21.
}
\usage{
MODEL_CONTEXT_LIMITS
}
\description{
Table of context window sizes for known models. Used by
[context_limit_for_model()]. Add new entries here as providers
ship them.
}
\keyword{internal}
