% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bump.R
\name{plot_bump}
\alias{plot_bump}
\title{Plot a bump chart / slope graph}
\usage{
plot_bump(df, time_var = "year", value_var = "value",
  region_var = "region", facet_var = NA, sort_by = "diff",
  sort_desc = TRUE, file_name = NA, width = 10, height = 6,
  saveBoth = FALSE, line_stroke = 0.5, dot_size = 5, dot_shape = 21,
  label_size = 4, label_x_offset = 0.3, value_y_offset = NA,
  label_vals = TRUE, percent_vals = FALSE, tufte_style = FALSE,
  x_buffer = 0.5, font_normal = "Lato", font_semi = "Lato",
  font_light = "Lato Light", panel_spacing = 3, font_axis_label = 12,
  font_axis_title = font_axis_label * 1.15, font_facet = font_axis_label *
  1.15, font_legend_title = font_axis_label,
  font_legend_label = font_axis_label * 0.8, font_subtitle = font_axis_label
  * 1.2, font_title = font_axis_label * 1.3, grey_background = FALSE,
  background_colour = grey10K, projector = FALSE)
}
\arguments{
\item{df}{Data frame}
}
\description{
Plot a bump chart / slope graph to show changes over time
}
\examples{
# generate random data
df = data.frame(year = c(rep(2007, 6), rep(2016, 6)), value = sample(1:100, 12), region = rep(letters[1:6], 2), facet = rep(c('group1', 'group2'), 6))
plot_bump(df, time_var = 'year', value_var = 'value', region_var = 'region', tufte_style = TRUE, facet_var = 'facet')
plot_bump(df, time_var = 'year', value_var = 'value', region_var = 'region', facet_var = 'region', sort_desc = FALSE, sort_by = 'first')
plot_bump(df, time_var = 'year', value_var = 'value', region_var = 'region', facet_var = 'region', sort_desc = FALSE, sort_by = 'last')
plot_bump(df, time_var = 'year', value_var = 'value', region_var = 'region', facet_var = 'region', sort_desc = FALSE, sort_by = 'diff')

}

