% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coef.R
\name{plot_coef}
\alias{plot_coef}
\title{Plot coefficients of a model with errors}
\usage{
plot_coef(model, negative_ontop = TRUE, negative_good = FALSE,
  cluster_col = NA, level = 0.95, CI_factor = 1.96, exclude_terms = NA,
  exclude_intercept = TRUE, plot_left_labels = TRUE,
  plot_right_labels = FALSE, alpha_insignificant = 0.3, size_point = 3,
  x_buffer = 0.1, label_margin = 1, font_normal = "Lato",
  font_semi = "Lato Light", font_light = "Lato Light")
}
\arguments{
\item{model}{Fitted model object from a lm or glm}

\item{negative_good}{Should negative coefficients be displayed in red (default) or blue (negative_good = TRUE)?}

\item{cluster_col}{Column in the *original* data frame used to build the model to be used to calculate clustered standards errors using the `multiwayvcov` package. By default, no clustering correction is used, and the standard erorrs are those calculated in the model.  If using clustered errors, should be specified as: 'orginal_data$clustering_column'}

\item{level}{confidence level for error bars / significance indicator}

\item{CI_factor}{Factor to adjust the standard errors by. By default, 1.96, assuming a normal distribution with sufficient sample size (95 percent of distribution lies within 1.96 standard deviations)}

\item{exclude_terms}{(optional) string of terms to exclude}

\item{exclude_intercept}{whether to include the intercept from the model in the color-coding of coefficients.}

\item{plot_left_labels}{include names of variables on the left side of the y-axis}

\item{plot_right_labels}{include names of variables on the right side of the y-axis}

\item{alpha_insignificant}{alpha (opacity) level for variables that are insignificant}

\item{size_point}{size to plot the coefficients, in mm}

\item{x_buffer}{percentage offset for the y-axis labels}

\item{label_margin}{offset for the y-axis labels}

\item{font_normal}{string containing the name of the font to use in darkest text}

\item{font_semi}{string containing the name of the font to use in medium dark text}

\item{font_light}{string containing the name of the font to use in lightest text}
}
\description{
Plots coefficients from a model, along with standard errors.  Coefficients that are statistically significant at agiven level are highlighted in darker text. Inspired by https://github.com/jaredlander/coefplot/blob/master/R/coefplot.r but works w/ ggplot2 version > 2.2
}
\examples{
data(diamonds, package = 'ggplot2')
model1 <- lm(price ~ carat + cut*color, data=diamonds)
plot_coef(model1)

# Sort of a contrived example to show the effect of clustering standard errors.
model2 <- lm(price ~ carat + color, data=diamonds)
# No clustered errors
plot_coef(model2)
# Errors clustered by cut
plot_coef(model2, cluster_col = diamonds$cut)
}
