% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{plot_color_palette}
\alias{plot_color_palette}
\title{Colors: internal colors based on RColorBrewer, Viridis, Vega and Tableau colors
From https://github.com/vega/vega/wiki/Scales}
\usage{
plot_color_palette(color_name)
}
\description{
Colors: internal colors based on RColorBrewer, Viridis, Vega and Tableau colors
From https://github.com/vega/vega/wiki/Scales
}
\examples{
# Sequential, warm tones
plot_color_palette('Reds')
plot_color_palette('OrRd')
plot_color_palette('Oranges')
plot_color_palette('YlOrRd')
plot_color_palette('YlOrBr')
plot_color_palette('RdPu')
plot_color_palette('PuRd')
plot_color_palette('magma')
plot_color_palette('inferno')
plot_color_palette('magma')
plot_color_palette('plasma')
# Sequential, cool tones
plot_color_palette('Greens')
plot_color_palette('YlGn')
plot_color_palette('YlGnBu')
plot_color_palette('viridis')
plot_color_palette('BuGn')
plot_color_palette('GnBu')
plot_color_palette('PuBuGn')
plot_color_palette('Blues')
plot_color_palette('Purples')
plot_color_palette('BuPu')
plot_color_palette('Greys')

# Diverging
plot_color_palette('RdYlGn')
plot_color_palette('RdYlBu')
plot_color_palette('RdBu')
plot_color_palette('RdGy')
plot_color_palette('Spectral')
plot_color_palette('PiYG')
plot_color_palette('PRGn')
plot_color_palette('PuOr')
plot_color_palette('BrBG')
# Categorical
plot_color_palette('category10')
plot_color_palette('category20')
plot_color_palette('category20b')
plot_color_palette('category20c')
plot_color_palette('tableau10')
plot_color_palette('tableau20')
plot_color_palette('Paired')
plot_color_palette('Pastel1')
plot_color_palette('Pastel2')
plot_color_palette('Accent')
plot_color_palette('Dark2')
plot_color_palette('Set1')
plot_color_palette('Set2')
plot_color_palette('Set3')
}
