% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{plot_corr}
\alias{plot_corr}
\title{Calculate and plot a correlation matrix.}
\usage{
plot_corr(df, use_arg = "pairwise.complete.obs", remove_same = TRUE,
  remove_half = TRUE, stroke_colour = "white", stroke_size = 0.5,
  square_tiles = TRUE, fill_palette = brewer.pal(11, "RdYlBu"),
  fill_limits = c(-1, 1), font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")
}
\arguments{
\item{df}{data frame containing only numeric values}
}
\description{
Plots coefficients from a model, along with standard errors.  Coefficients that are statistically significant at agiven level are highlighted in darker text. Inspired by https://github.com/jaredlander/coefplot/blob/master/R/coefplot.r but works w/ ggplot2 version > 2.2
}
\examples{
plot_corr(mtcars)
}

