% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.R
\name{plot_dotplot}
\alias{plot_dotplot}
\title{Plots a dot plot}
\usage{
plot_dotplot(df, time_var = "year", region_var = "region",
  value_var = "avg", sort_desc = TRUE, sort_by = "diff", facet_var = NA,
  ncol = NULL, nrow = NULL, scales = "fixed", include_arrows = TRUE,
  arrow_arg = arrow(length = unit(0.03, "npc")), connector_length = 0.85,
  dot_size = 6, dot_shape = c(21, 23, 22, 24), fill_value = TRUE,
  dot_fill_discrete = c("#D3DEED", "#3288BD"), dot_fill_cont = brewer.pal(9,
  "YlGnBu"), connector_stroke = 0.25, connector_colour = grey75K,
  label_vals = TRUE, label_size = 3, label_colour = grey75K,
  label_digits = 1, percent_vals = FALSE, value_y_offset = 0,
  horiz = TRUE, file_name = NA, width = 10, height = 6,
  saveBoth = FALSE, font_normal = "Lato", font_semi = "Lato",
  font_light = "Lato Light", panel_spacing = 1, font_axis_label = 12,
  font_axis_title = font_axis_label * 1.15, font_facet = font_axis_label *
  1.15, font_legend_title = font_axis_label,
  font_legend_label = font_axis_label * 0.8, font_subtitle = font_axis_label
  * 1.2, font_title = font_axis_label * 1.3, grey_background = FALSE,
  background_colour = grey10K, projector = FALSE)
}
\description{
Plots a dot plot
}
\examples{
# generate random data
df = data.frame(year = c(rep(2007, 6), rep(2016, 6)), value = sample(1:100, 12), region = rep(letters[1:6], 2), facet = rep(c('group1', 'group2'), 6))

plot_dotplot(df, time_var = 'year', region_var = 'region', value_var = 'value')
plot_dotplot(df, time_var = 'year', region_var = 'region', value_var = 'value', include_arrows = FALSE)
plot_dotplot(df, time_var = 'year', region_var = 'region', value_var = 'value', sort_by = 'first', fill_value = FALSE, value_y_offset = 0.25, sort_desc = FALSE)
plot_dotplot(df, time_var = 'year', region_var = 'region', value_var = 'value', sort_by = 'first', fill_value = FALSE, value_y_offset = 0.25, sort_desc = FALSE)
 
}

