% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBasicMap.R
\name{plotMap}
\alias{plotMap}
\title{Import ESRI shapefiles into R}
\usage{
plotMap(df, exportPlot = FALSE, fileName = "map.pdf", plotWidth = 6,
  plotHeight = 6)
}
\description{
Import shapefile and convert into lat/lon coordinates to plot in R or Tableau
}
\details{
View available projections with 'projInfo(type = 'proj')'
View available datum with 'projInfo(type = 'datum')'
View available ellipsoids with 'projInfo(type = 'ellps')'
}
\examples{



}
\author{
Laura Hughes, laura.d.hughes@gmail.com
}

