% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recoding and cleaning of survey data}
\usage{
recode(df, dict, oldVar, newVar)
}
\arguments{
\item{df}{data frame to change}

\item{dict}{Assumes that the dictionary will be a 2-column data frame with key in first column and value in the second. Third column is an optional column explaining the relationship.  Should be some sort of logical operator, e.g. '==', '<', '<=', ...}

\item{oldVar}{name of the variable to change}

\item{newVar}{name of the new variable to generate}
}
\description{
Converts values of a data frame to encoded values using a dictionary.
}

