% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importShp.R
\name{shp2csv}
\alias{shp2csv}
\title{Import ESRI shapefiles into R}
\usage{
shp2csv(workingDir = getwd(), layerName, exportData = TRUE,
  fileName = layerName, labelVar = NA, reproject = TRUE,
  projection = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0")
}
\arguments{
\item{workingDir}{string containing the name of the folder containing the shapefile}

\item{layerName}{string containing the name of the shapefile, e.g. 'district_boundary' for 'district_boundary.shp'}

\item{exportData}{Boolean for whether to save the data as a .csv after importing}

\item{fileName}{string containing the name of the .csv to be exported; by default, will be the same as the inputted .shp file}

\item{labelVar}{string containing the variable name containing the names of the polygons (for instance, the names of the provinces, districts, etc.)}

\item{reproject}{Boolean specifying whether to reproject the data}

\item{projection}{CRS projection string to standardize the projection of the shapefile}
}
\description{
Import shapefile and convert into lat/lon coordinates to plot in R or Tableau
}
\details{
View available projections with 'projInfo(type = 'proj')'
View available datum with 'projInfo(type = 'datum')'
View available ellipsoids with 'projInfo(type = 'ellps')'
}
\examples{



}
\author{
Laura Hughes, laura.d.hughes@gmail.com
}

