% tinyrox says don't edit this manually, but it can't stop you!
\name{task_delete}
\alias{task_delete}
\title{Delete a task}
\usage{
task_delete(task_id, agent_id = "main")
}
\arguments{
\item{task_id}{Task ID}

\item{agent_id}{Agent identifier}
}
\value{
Invisible TRUE on success
}
\description{
Delete a task
}
