% tinyrox says don't edit this manually, but it can't stop you!
\name{task_get}
\alias{task_get}
\title{Get a task by ID}
\usage{
task_get(task_id, agent_id = "main")
}
\arguments{
\item{task_id}{Task ID}

\item{agent_id}{Agent identifier}
}
\value{
Task list, or NULL if not found
}
\description{
Get a task by ID
}
