% tinyrox says don't edit this manually, but it can't stop you!
\name{task_resume}
\alias{task_resume}
\title{Resume a paused task}
\usage{
task_resume(task_id, agent_id = "main")
}
\arguments{
\item{task_id}{Task ID}

\item{agent_id}{Agent identifier}
}
\value{
Invisible TRUE
}
\description{
Resume a paused task
}
