% tinyrox says don't edit this manually, but it can't stop you!
\name{task_runs}
\alias{task_runs}
\title{Get runs for a task}
\usage{
task_runs(task_id, limit = 10L, agent_id = "main")
}
\arguments{
\item{task_id}{Task ID}

\item{limit}{Max runs to return (default: 10)}

\item{agent_id}{Agent identifier}
}
\value{
Data frame of runs
}
\description{
Get runs for a task
}
