% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_xylab}
\alias{theme_xylab}
\title{Theme with x and y axis labels}
\usage{
theme_xylab()
}
\description{
This function can be tacked onto a ggplot object to specify a new theme with the x and y axis labeled.
}

