% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{themes}
\alias{theme_basic}
\alias{theme_blank}
\alias{theme_labelsOnly}
\alias{theme_legend}
\alias{theme_stroke}
\alias{theme_xaxis}
\alias{theme_xgrid}
\alias{theme_xygrid}
\alias{theme_xygridlight}
\alias{theme_xylab}
\alias{theme_yaxis}
\alias{theme_ygrid}
\alias{theme_ylab}
\alias{themes}
\title{Customized ggplot themes}
\usage{
theme_basic(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light", large_font = 15, medium_font = 14,
  small_font = 12)

theme_xygrid(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_xygridlight(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_xgrid(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light", legend.position = "none",
  legend.direction = "horizontal", font_axis_label = 12,
  font_axis_title = font_axis_label * 1.15, font_facet = font_axis_label *
  1.15, font_legend_title = font_axis_label,
  font_legend_label = font_axis_label * 0.8, font_title = font_axis_label *
  1.3, grey_background = FALSE, background_colour = grey10K,
  projector = FALSE)

theme_stroke(stroke_size = 0.25, stroke_colour = grey90K)

theme_ygrid(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_blank(legend.position = "none")

theme_legend(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_xaxis(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_yaxis(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_xylab(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_ylab(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_labelsOnly(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")
}
\arguments{
\item{font_normal}{string containing font name for normal text}

\item{font_semi}{string containing font name for semilight text}

\item{font_light}{string containing font name for light text}

\item{legend.position}{(optional) legend position; takes either a ggplot legend position string like `'left'` or a tuple in percent of the x- and y-axes, like `c(0.3, 0.7)`}

\item{projector}{(optional) if TRUE, make lines and text bolder for an LCD projector}

\item{font_size}{value, in points, of the font size}
}
\description{
Customized ggplot themes
}
\section{Functions}{
\itemize{
\item \code{theme_basic}: Theme with axis labels, titles, grid lines, axes, and legend.

\item \code{theme_xygrid}: Theme with x and y labels, titles, and gridlines

\item \code{theme_xygridlight}: Theme with x and y labels, titles, and gridlines

\item \code{theme_xgrid}: Theme with light x-grid lines, x and y axis labels, and x-axis title.

\item \code{theme_stroke}: Theme with thin grey border around edge

\item \code{theme_ygrid}: Theme with light y-grid lines, x and y axis labels, and y-axis title.

\item \code{theme_blank}: completely blank theme; similar to theme_void but without legend or margins.

\item \code{theme_legend}: completely blank theme; similar to theme_void but without legend or margins.

\item \code{theme_xaxis}: Theme with x labels, x title, and x-axis line; no gridlines

\item \code{theme_yaxis}: Theme with y labels, titles, and y-axis line; no gridlines

\item \code{theme_xylab}: Theme with x and y axis labels; no gridlines.

\item \code{theme_ylab}: Theme with  y axis labels; no gridlines.

\item \code{theme_labelsOnly}: Theme with only x and y labels; no axes, axis titles, gridlines, or legends
}}
\examples{
ggplot(mtcars, aes(x = wt, y = mpg, colour = cyl)) + geom_point() + facet_wrap(~am) + ggtitle('Heavier cars have worse gas efficiency') + theme_xgrid(legend.position = c(0.8, 0.9), grey_background = T)

}

