% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPtEst.R
\name{calcPtEst}
\alias{calcPtEst}
\title{Calculate point estimates and errors}
\usage{
calcPtEst(df, var, use_weights = FALSE, by_var = NA, design = NA,
  psu_var = NA, strata_var = NA, weight_var = NA, na.rm = TRUE,
  ci_factor = 2)
}
\arguments{
\item{df}{main data frame containing the raw data}

\item{var}{string containing the variable name you want to average}

\item{use_weights}{TRUE/FALSE for whether to apply a sampling frame or calculate a simple average.}

\item{by_var}{(optional) string containing the variable name over which you want to calculate the estimates}

\item{design}{(optional) svydesign object containing the sample frame}

\item{psu_var}{(optional) if design isn't specified, string containing the primary sampling unit variable from the survey design (argument \code{id} in library('survey'))}

\item{strata_var}{(optional) if design isn't specified, string containing the strata variable from the survey design (argument \code{strata} in library('survey')}

\item{weight_var}{(optional)if design isn't specified, string containing the weights variable from the survey design (argument \code{weights} in library('survey')}

\item{na.rm}{remove NAs from the mean or not}

\item{ci_factor}{value to calculate confidence interval; in standard deviations. 1.96 standard deviations --> ~ 95 percent of the area under a normal gaussian distribution}
}
\description{
Calculates point estimates and errors using the package `survey` based on sampling weights from a survey design
}
\examples{
# Generate simple random data to average
df = data.frame(region = rep(letters, 100), value = sample(100, 2600, replace = TRUE))

calcPtEst(df, 'value', by_var = 'region', use_weights = FALSE)

# generate data from a survey
data(api, package = 'survey')
dclus1<-svydesign(id=~dnum, weights=~pw, data=apiclus1, fpc=~fpc)
# no weights
calcPtEst(apiclus1, var = 'api99', by_var = 'stype', use_weights = FALSE)
# specifying the sampling frame as a design object
calcPtEst(apiclus1, var = 'api99', by_var = 'stype', design = dclus1)

}

