% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.R
\name{coefplot}
\alias{coefplot}
\title{Plot coefficients of a model with errors}
\usage{
coefplot(model, negative_good = FALSE, level = 0.95,
  exclude_intercept = TRUE, plot_left_labels = TRUE,
  plot_right_labels = FALSE, alpha_insignificant = 0.3, size_point = 3,
  x_buffer = 0.1, label_margin = 0.4, font_normal = "Lato",
  font_semi = "Lato Light", font_light = "Lato Light")
}
\arguments{
\item{model}{Fitted model}
}
\description{
Inspired by https://github.com/jaredlander/coefplot/blob/master/R/coefplot.r
but works w/ ggplot2 version > 2.2
}
\examples{
data(diamonds)
model1 <- lm(price ~ carat + cut*color, data=diamonds)
coefplot(model1)

}

