% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatNum.R
\name{formatNum}
\alias{formatNum}
\alias{round_mean}
\alias{round_std}
\alias{percent}
\alias{round_exact}
\title{Simple reusable functions to format numbers}
\usage{
round_mean(x, ndigits = 2)

round_std(x, ndigits = 2)

percent(x, ndigits = 1)

round_exact(x, ndigits = 1)
}
\arguments{
\item{x}{numbers to be converted}

\item{ndigits}{number of digits to which to round the number}
}
\description{
Simple reusable functions to format numbers
}
\section{Functions}{
\itemize{
\item \code{round_mean}: Function to calculate the mean and round to two digits.

\item \code{round_std}: Function to calculate the standard deviation and round to two digits.

\item \code{percent}: Function to convert a decimal to a percent, round to a specific number of digits, and convert to a string with a percent sign.

\item \code{round_exact}: Function to round a number w/ an exact number of digits. Returns a string.
}}

\examples{
round_mean(mtcars$mpg)
round_exact(2, ndigits = 3)
round_std(mtcars$mpg)
percent(0.157389)
}
