% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_get_verbosity}
\alias{llama_get_verbosity}
\title{Get current verbosity level}
\usage{
llama_get_verbosity()
}
\value{
Integer verbosity level (0-3)
}
\description{
Get current verbosity level
}
\examples{
llama_get_verbosity()
}
