% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_perf_reset}
\alias{llama_perf_reset}
\title{Reset performance counters}
\usage{
llama_perf_reset(ctx)
}
\arguments{
\item{ctx}{Context handle returned by [llama_new_context]}
}
\value{
No return value, called for side effects.
}
\description{
Resets the timing and token count statistics for the context.
}
\examples{
if (FALSE) {
llama_perf_reset(ctx)
}
}
