% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_supports_mmap}
\alias{llama_supports_mmap}
\title{Check whether memory-mapped file I/O is supported}
\usage{
llama_supports_mmap()
}
\value{
A logical scalar: \code{TRUE} if mmap is supported.
}
\description{
Check whether memory-mapped file I/O is supported
}
\examples{
llama_supports_mmap()
}
