% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{plot_dot}
\alias{plot_dot}
\title{Plot a dot plot, either with or without axis bars}
\usage{
plot_dot(df, by_var = "region", value_var = "avg", use_weights = TRUE,
  sort_desc = TRUE, sort_by = "diff", reference_line = TRUE,
  line_stroke = 0.25, line_colour = grey75K, lollipop = FALSE,
  facet_var = NA, ncol = NULL, nrow = NULL, scales = "fixed",
  include_arrows = TRUE, arrow_arg = arrow(length = unit(0.03, "npc")),
  connector_length = 0.85, dot_size = 6, dot_shape = 21,
  fill_value = TRUE, dot_fill_discrete = c("#D3DEED", "#3288BD"),
  dot_fill_cont = brewer.pal(9, "YlGnBu"), connector_stroke = 0.25,
  connector_colour = grey75K, label_vals = TRUE, label_size = 3,
  label_colour = grey75K, label_digits = 1, percent_vals = FALSE,
  value_label_offset = 0, label_group = TRUE, label_group_size = 4,
  group_label_offset = 0.25, horiz = TRUE, file_name = NA, width = 10,
  height = 6, saveBoth = FALSE, font_normal = "Lato",
  font_semi = "Lato", font_light = "Lato Light", panel_spacing = 1,
  font_axis_label = 12, font_axis_title = font_axis_label * 1.15,
  font_facet = font_axis_label * 1.15, font_legend_title = font_axis_label,
  font_legend_label = font_axis_label * 0.8, font_subtitle = font_axis_label
  * 1.2, font_title = font_axis_label * 1.3, legend.position = "none",
  legend.direction = "horizontal", grey_background = FALSE,
  background_colour = grey10K, projector = FALSE)
}
\description{
Plot a dot plot, either with or without axis bars
}
\examples{
# generate random data
df = data.frame(value = sample(1:100, 8), region = letters[1:8])

plot_dot(df, by_var = 'region', value_var = 'value')

}

