% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.R
\name{plot_dotplot}
\alias{plot_dotplot}
\title{Plots a dot plot}
\usage{
plot_dotplot(df, time_var = "year", region_var = "region", facet_var = NA,
  value_var = "avg", sort_desc = TRUE, arrow_arg = arrow(length =
  unit(0.03, "npc")), dot_size = 6, connector_stroke = 0.25,
  connector_colour = grey75K, label_vals = TRUE, label_colour = grey75K,
  percent_vals = FALSE, value_y_offset = 0.05, sort_by = "last",
  horiz = TRUE, file_name = NA, width = 10, height = 6,
  saveBoth = FALSE, font_normal = "Lato", font_semi = "Lato",
  font_light = "Lato Light", panel_spacing = 1, font_axis_label = 12,
  font_axis_title = font_axis_label * 1.15, font_facet = font_axis_label *
  1.15, font_legend_title = font_axis_label,
  font_legend_label = font_axis_label * 0.8, font_subtitle = font_axis_label
  * 1.2, font_title = font_axis_label * 1.3, grey_background = FALSE,
  background_colour = grey10K, projector = FALSE)
}
\description{
Plots a dot plot
}

