% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Simple wrapper to ggsave}
\usage{
save_plot(filename, plot = last_plot(), width = NA, height = NA,
  units = "in", scale = 1)
}
\description{
Wrapper to ggsave to auto-save the annoying fiddly arguments I always forget
}

