% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{themes}
\alias{theme_basic}
\alias{theme_blank}
\alias{theme_labelsOnly}
\alias{theme_legend}
\alias{theme_xaxis}
\alias{theme_xgrid}
\alias{theme_xygrid}
\alias{theme_xygridlight}
\alias{theme_xylab}
\alias{theme_yaxis}
\alias{theme_ygrid}
\alias{themes}
\title{Customized ggplot themes}
\usage{
theme_basic(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_xygrid(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_xygridlight(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_xgrid(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_ygrid(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_blank()

theme_legend(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_xaxis(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_yaxis(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")

theme_xylab()

theme_labelsOnly(font_normal = "Lato", font_semi = "Lato Light",
  font_light = "Lato Light")
}
\arguments{
\item{font_normal}{string containing font name for normal text}

\item{font_semi}{string containing font name for semilight text}

\item{font_light}{string containing font name for light text}
}
\description{
Customized ggplot themes
}
\section{Functions}{
\itemize{
\item \code{theme_basic}: Theme with axis labels, titles, grid lines, axes, and legend.

\item \code{theme_xygrid}: Theme with x and y labels, titles, and gridlines

\item \code{theme_xygridlight}: Theme with x and y labels, titles, and gridlines

\item \code{theme_xgrid}: Theme with light x-grid lines, x and y axis labels, and x-axis title.

\item \code{theme_ygrid}: Theme with light y-grid lines, x and y axis labels, and y-axis title.

\item \code{theme_blank}: completely blank theme; similar to theme_void but without legend or margins.

\item \code{theme_legend}: completely blank theme; similar to theme_void but without legend or margins.

\item \code{theme_xaxis}: Theme with x labels, x title, and x-axis line; no gridlines

\item \code{theme_yaxis}: Theme with y labels, titles, and y-axis line; no gridlines

\item \code{theme_xylab}: Theme with x and y axis labels; no gridlines.

\item \code{theme_labelsOnly}: Theme with only x and y labels; no axes, axis titles, gridlines, or legends
}}
\examples{
ggplot(mtcars, aes(x = mpg, y = wt, colour = cyl)) + geom_point() + theme_xylab()

}

